Workshop participants will not have `make` or `sphinx-build`
The `npm install && npm start` workflow is for workshop participants.  Participants will be working with a released version of the workshop package, which includes the build artifacts in the `dist` directory.  Workshop developers are responsible for generating these build artifacts.  Workshop developers need to have `make` and `sphinx-build` installed.  Participants will follow the setup instructions in `README.md` and developers follow the instructions in `DEV-README.md`.